Exclude HTMLUnit failing assertion in Jenkins 2.400 and later
https://github.com/jenkinsci/bom/pull/1959#issuecomment-1507906453
explains that the modernization of hetero-list.js has inserted JavaScript
that the HTMLUnit JavaScript engine does not understand.  That causes
the test to fail with an HTTP 500 error when the specific assertion
is executed.

Do not execute the failing assertion on Jenkins 2.400 and
later.  Excluding the failing assertion is a simple short-term
workaround for the larger problem that is described by @basil in
https://github.com/jenkinsci/jenkins-test-harness/issues/569#issue-1661461679