allow markdown and html files for documentation of library steps
the current approach with the txt files has the problem that the content
must fit the formatter that is globally configured. When a pipeline
library is generic enough to be used on many Jenkins instances the
documentation might be broken.
As Markdown allows to easily write the docs this should be the new
default.
The snippet is wrapped in a div so we can apply a different styling. The
default Jenkins styling doesn't look so good for markdown files at least
(no margin before `<hx>` and big margin afterwards)