Move transitive dependency version declaration to dependencyManagement
The commons-lang3-api plugin dependency is not a direct dependency of
the plugin (either in compile scope or test scope), but is needed by
one of the test dependencies.

Other temporary version declarations avoid duplicating the version
declaration at multiple locations in the pom and make it simpler to
remove them after a future plugin BOM update.